+++ dask_geomodeling/tests/test_geometry_sinks.py
@@ -14,6 +14,7 @@ from dask_geomodeling.tests.factories import (
     setup_temp_root,
     teardown_temp_root,
 )
+
 try:
     from pandas.testing import assert_frame_equal
 except ImportError:
@@ -25,6 +26,14 @@ def compare_crs(actual, expected):
         assert actual == expected
     else:
         assert actual["init"] == expected["init"]
+
+
+def assert_frame_equal_ignore_index(actual, expected, sort_col):
+    assert_frame_equal(
+        actual.set_index(sort_col).sort_index(),
+        expected.set_index(sort_col).sort_index(),
+        check_like=True,
+    )
 
 
 class TestGeometryFileSink(unittest.TestCase):
@@ -52,7 +61,7 @@ class TestGeometryFileSink(unittest.TestCase):
         self.request_tiled = {
             "mode": "centroid",
             "projection": "EPSG:3857",
-            "geometry": box(0, 0, 20, 20)
+            "geometry": box(0, 0, 20, 20),
         }
         self.path = os.path.join(self.root, self._testMethodName)
         self.polygons = [
@@ -110,7 +119,7 @@ class TestGeometryFileSink(unittest.TestCase):
         expected = self.expected_to_geojson(self.expected)
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, expected, "int")
         # compare projections
         compare_crs(actual.crs, expected.crs)
 
@@ -126,7 +135,7 @@ class TestGeometryFileSink(unittest.TestCase):
         actual = gpd.read_file(os.path.join(self.path, filename))
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, self.expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, self.expected, "int")
         # compare projections
         compare_crs(actual.crs, self.expected.crs)
 
@@ -138,7 +147,7 @@ class TestGeometryFileSink(unittest.TestCase):
         actual = gpd.read_file(os.path.join(self.path, filename))
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, self.expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, self.expected, "int")
         # compare projections
         compare_crs(actual.crs, self.expected.crs)
 
@@ -157,7 +166,7 @@ class TestGeometryFileSink(unittest.TestCase):
         del actual["fid"]
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, self.expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, self.expected, "int")
 
     def test_fields_non_available(self):
         with pytest.raises(ValueError):
@@ -188,7 +197,7 @@ class TestGeometryFileSink(unittest.TestCase):
         expected = self.expected_to_geojson(self.expected_combined)
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, expected, "int")
         # compare projections
         compare_crs(actual.crs, expected.crs)
 
@@ -221,20 +230,14 @@ class TestGeometryFileSink(unittest.TestCase):
 
     def test_categorical_column(self):
         with_categorical = self.source.set(
-            "categorical",
-            Classify(self.source["float"], bins=[6], labels=["A", "B"])
+            "categorical", Classify(self.source["float"], bins=[6], labels=["A", "B"])
         )
         block = self.klass(
-            with_categorical,
-            self.path,
-            "geojson",
-            fields={"label": "categorical"},
+            with_categorical, self.path, "geojson", fields={"label": "categorical"}
         )
         block.get_data(**self.request)
 
-        actual = gpd.read_file(
-            os.path.join(self.path, os.listdir(self.path)[0])
-        )
+        actual = gpd.read_file(os.path.join(self.path, os.listdir(self.path)[0]))
         assert actual["label"].tolist() == ["A"]
 
     def test_to_file_geojson(self):
@@ -244,32 +247,26 @@ class TestGeometryFileSink(unittest.TestCase):
         expected = self.expected_to_geojson(self.expected)
 
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, expected, "int")
 
     def test_to_file_shapefile(self):
         self.source.to_file(self.path + ".shp", **self.request)
         actual = gpd.read_file(self.path + ".shp")
         # compare dataframes without checking the order of records / columns
-        assert_frame_equal(actual, self.expected, check_like=True)
+        assert_frame_equal_ignore_index(actual, self.expected, "int")
 
     def test_to_file_with_tiling_geojson(self):
-        self.source.to_file(
-            self.path + ".geojson", tile_size=10, **self.request_tiled
-        )
+        self.source.to_file(self.path + ".geojson", tile_size=10, **self.request_tiled)
         actual = gpd.read_file(self.path + ".geojson")
         # because we lose the index in the saving process, just check the len
         assert len(actual) == 2
 
     def test_to_file_dry_run(self):
-        self.source.to_file(
-            self.path + ".geojson", dry_run=True, **self.request
-        )
+        self.source.to_file(self.path + ".geojson", dry_run=True, **self.request)
         assert not os.path.exists(self.path)
 
     def test_to_file_with_tiling_shapefile(self):
-        self.source.to_file(
-            self.path + ".shp", tile_size=10, **self.request_tiled
-        )
+        self.source.to_file(self.path + ".shp", tile_size=10, **self.request_tiled)
         actual = gpd.read_file(self.path + ".shp")
         # because we lose the index in the saving process, just check the len
         assert len(actual) == 2